+++ tests/api/backup.rs
@@ -94,7 +94,7 @@ pub fn simple_backup_with_excludes() -> Result<()> {
     // TODO: Check index stats.
     // TODO: Check what was restored.
 
-    let validate_stats = af.validate().unwrap();
+    let validate_stats = af.validate(&ValidateOptions::default()).unwrap();
     assert!(!validate_stats.has_problems());
     Ok(())
 }