+++ tests/api/damaged.rs
@@ -21,7 +21,19 @@ use conserve::*;
 fn missing_block() -> Result<()> {
     let archive = Archive::open_path(Path::new("testdata/damaged/missing-block"))?;
 
-    let validate_stats = archive.validate()?;
+    let validate_stats = archive.validate(&ValidateOptions::default())?;
+    assert_eq!(validate_stats.has_problems(), true);
+    assert_eq!(validate_stats.block_missing_count, 1);
+    Ok(())
+}
+
+#[test]
+fn missing_block_skip_block_hashes() -> Result<()> {
+    let archive = Archive::open_path(Path::new("testdata/damaged/missing-block"))?;
+
+    let validate_stats = archive.validate(&ValidateOptions {
+        skip_block_hashes: true,
+    })?;
     assert_eq!(validate_stats.has_problems(), true);
     assert_eq!(validate_stats.block_missing_count, 1);
     Ok(())